==> notebooks/1_quality.ipynb <==
# Installing required packages
!pip install pyspark
!pip install findspark

# starting spark
import findspark
findspark.init()

# PySpark is the Spark API for Python. In this lab, we use PySpark to initialize the spark context. 
from pyspark import SparkContext, SparkConf
from pyspark.sql import SparkSession
from pyspark.sql.functions import when, col, lit, udf, length, substring, expr, regexp_replace, sum_distinct
from pyspark.sql.types import StructType,StructField, StringType, IntegerType, FloatType
from pyspark.sql import functions as F

# Creating a spark context class
sc = SparkContext()

# Creating a spark session
spark = (SparkSession.builder
                     .master("local[7]")
                     .appName("Aceleração PySpark - Capgemini"))

schema_airports = StructType([
    StructField("faa",  StringType(),  True),
    StructField("name", StringType(),  True),
    StructField("lat",  FloatType(),   True),
    StructField("lon",  FloatType(),   True),
    StructField("alt",  IntegerType(), True),
    StructField("tz",   IntegerType(), True),
    StructField("dst",  StringType(),  True)
])

schema_planes = StructType([
    StructField("tailnum",      StringType(),  True),
    StructField("year",         IntegerType(), True),
    StructField("type",         StringType(),  True),
    StructField("manufacturer", StringType(),  True),
    StructField("model",        StringType(),  True),
    StructField("engines",      IntegerType(), True),
    StructField("seats",        IntegerType(), True),
    StructField("speed",        IntegerType(), True),
    StructField("engine",       StringType(),  True)
])

schema_flights = StructType([
    StructField("year",      IntegerType(), True),
    StructField("month",     IntegerType(), True),
    StructField("day",       IntegerType(), True),
    StructField("dep_time",  StringType(),  True),
    StructField("dep_delay", IntegerType(), True),
    StructField("arr_time",  StringType(),  True),
    StructField("arr_delay", IntegerType(), True),
    StructField("carrier",   StringType(),  True),
    StructField("tailnum",   StringType(),  True),
    StructField("flight",    StringType(),  True),
    StructField("origin",    StringType(),  True),
    StructField("dest",      StringType(),  True),
    StructField("air_time",  IntegerType(), True),
    StructField("distance",  IntegerType(), True),
    StructField("hour",      IntegerType(), True),
    StructField("minute",    IntegerType(), True),
])

df_airports = (spark.getOrCreate().read
                  .format("csv")
                  .option("header", "true")
                  .schema(schema_airports)
                  .load("../Data/airports.csv"))

df_planes = (spark.getOrCreate().read
                  .format("csv")
                  .option("header", "true")
                  .schema(schema_planes)
                  .load("../Data/planes.csv"))

df_flights = (spark.getOrCreate().read
                  .format("csv")
                  .option("header", "true")
                  .schema(schema_flights)
                  .load("../Data/flights.csv"))

# Expressoes regulares comuns
REGEX_ALPHA    = r'[a-zA-Z]+'
REGEX_INTEGER  = r'[0-9]+'
REGEX_FLOAT    = r'[0-9]+\.[0-9]+'
REGEX_ALPHANUM = r'[0-9a-zA-Z]+'
REGEX_EMPTY_STR= r'[\t ]+$'
REGEX_SPECIAL  = r'[!@#$%&*\(\)_]+'
REGEX_NNUMBER  = r'^N[1-9][0-9]{2,3}([ABCDEFGHJKLMNOPQRSTUVXWYZ]{1,2})'
REGEX_NNUMBER_INVALID = r'(N0.*$)|(.*[IO].*)'
REGEX_TIME_FMT = r'^(([0-1]?[0-9])|(2[0-3]))([0-5][0-9])$'

# Funcoes auxiiliares
def split_csv(line):
    return tuple(map(lambda x: x.replace('"',''), line.split(",")))

def check_empty_column(col):
    return (F.col(col).isNull() | (F.col(col) == '') | F.col(col).rlike(REGEX_EMPTY_STR))

# Airports

# Pergunta 1

df1 = df_airports.withColumn(
    'qa_faa', 
    when(
            (
                (col('faa').isNull()) |
                (col('faa') == "NA")
            ), "M"
        )
    .when(~F.col("faa").rlike("^(?=.*[0-9])(?=.*[a-zA-Z])(?=[A-Z0-9]).{3,5}$"), "F")
)

# Teste do resultado

df_airports.select('faa', 'qa_faa').where(col('qa_faa')=='F').show()

df1.groupBy('qa_faa').count().show()

### Pergunta 1 revisada 
- Utilizando o mesmo nome do Dataframe<br> 
- Utilizando função para realizar o tratamento

def check_empty_column(col):
    return (F.col(col).isNull() | (F.col(col) =='') | F.col(col).rlike(REGEX_EMPTY_STR))

df_airports = df_airports.withColumn('qa_faa',(
        F.when(check_empty_column('faa'),'M')
         .when(
             (~F.length(F.col('faa')).between(3,5)  |


df_airports.select('faa', 'qa_faa').where(col('qa_faa')=='F').show()

# Pergunta 2

df2 = df1.withColumn(
    'qa_name', 
    when(
            (
                (col('name').isNull()) |
                (col('name') == "NA")
            ), "M"
        )
)

## Pergunta 2 revisada 
- Utilizando função para tratar valor nulo e na

# Usando dataframe
df_airports = df_airports.withColumn('qa_name', F.when(check_empty_column('name'), 'M'))
df_airports.show(5)

# Teste do resultado

df_airports.groupBy('qa_name').count().show()

df_airports.select('name', 'qa_name').where(col('qa_name')=='M').show()

# Pergunta 3

df3 = df2.withColumn(
    "qa_lat",
    when(~col("lat").between(-180, 180), "I")
    .when(
            (
                (col('lat').isNull()) |
                (col('lat') == "NA")
            ), "M"
        )
    .when(col("lat").rlike(REGEX_ALPHA), "A"))

### pergunta 3 revisada
- O código está igual, só a formatação que está direferente da minha. 

# Usando Dataframe
df_airports = df_airports.withColumn('qa_lat',(
      F.when(check_empty_column('lat'), 'M')
       .when(F.col('lat').rlike(REGEX_ALPHA),'A')
       .when(F.col('lat').between(-180,180),'I')))

df_airports.show(5)

# Teste do resultado

df_airports.groupBy('qa_lat').count().show()

# Pergunta 4

df4 = df3.withColumn(
    "qa_lon",
    when(~col("lon").between(-180, 180), "I")
    .when(
            (
                (col('lon').isNull()) |
                (col('lon') == "NA")
            ), "M"
        )
    .when(col("lon").rlike(REGEX_ALPHA), "A"))

# Pergunta 4 revisada
- Código igual apenas a formatação diferente.

# Usando dataframe
df_airports = df_airports.withColumn("qa_lon", (
    F.when(check_empty_column("lon"), "M")
     .when(~F.col('lon').rlike(REGEX_FLOAT),"A")
     .when(~F.col('lon').between(-180,180),"I")))

df_airports.show(5)

# Teste do resultado

df_airports.groupBy('qa_lon').count().show()

# Pergunta 5

df5 = df4.withColumn(
    "qa_alt",
    when(col("alt") < 0, "I")
    .when(
            (
                (col('alt').isNull()) |
                (col('alt') == "NA")
            ), "M"
          )
    .when(~col("alt").rlike("[0-9]+"), "A")
)

### Pergunta 5 revisada
- código o mesmo apenas a formatação diferente.

# Usando dataframe
df_airports = df_airports.withColumn('qa_alt', (
            F.when(check_empty_column('alt'), 'M')
             .when(~F.col('alt').rlike(REGEX_INTEGER), 'A')             
             .when(F.col('alt') < 0, 'I')))

df_airports.show(5)

# Teste do resultado

df_airports.groupBy(col('qa_alt')).count().show()

# Pergunta 6

df6 = df5.withColumn(
    "qa_tz",
    when(
            (
                (col('tz').isNull()) |
                (col('tz') == "NA")
            ), "M"
        )
    .when(~col('tz').between(-11, 14), "I")
    .when(~F.col('tz').rlike(REGEX_INTEGER + '$'), 'A')
)


#  Pergunta 6 revisada
- código o mesmo formatacao diferente

# Usando dataframe
df_airports = df_airports.withColumn('qa_tz', (
            F.when(check_empty_column('tz'), 'M')
             .when(~F.col('tz').rlike(REGEX_INTEGER + '$'), 'A')         
             .when(~F.col('tz').between(-11, 14), 'I')))

df_airports.show(5)

# Teste do resultado

df_airports.groupBy(col('qa_tz')).count().show()

# Pergunta 7

DST_CATEGORIES = ['E', 'A', 'S', 'O', 'Z', 'N', 'U']

df7 = df6.withColumn(
    "qa_dst",
    when(
            (
                (col('dst').isNull()) |
                (col('dst') == "NA")
            ), "M"
        )
             .when(F.col('dst').rlike(REGEX_INTEGER), 'N')
             .when(~F.col('dst').isin(DST_CATEGORIES), 'C')
)

# Pergunta 7 revisada
- código o mesmo, formatação diferente.

DST_CATEGORIES = ['E', 'A', 'S', 'O', 'Z', 'N', 'U']

# Usando dataframe
df_airports = df_airports.withColumn('qa_dst', (
            F.when(check_empty_column('dst'), 'M')
             .when(F.col('dst').rlike(REGEX_INTEGER), 'N')
             .when(~F.col('dst').isin(DST_CATEGORIES), 'C')))

df_airports.show(5)

# Teste do resultado

df_airports.groupBy(col('qa_dst')).count().show()

# Salvando o arquivo em parquet com as colunas qa e informações importantes

df7.select(
    col('faa'), 
    col('name'), 
    col('qa_faa'), 
    col('qa_name'), 
    col('qa_lat'), 
    col('qa_lon'), 
    col('qa_alt'), 
    col('qa_tz'), 
    col('qa_dst')
).write.parquet(
    path = 'C:/Users/coskata/Downloads/Datasets/parquet/airports.parquet',
    mode = 'overwrite'
)

# Salvando de um modo diferente sem passar todas as colunas.

(df_airports.select(df_airports.colRegex("`^qa_.*`"))
            .repartition(1) # coalesce
            .write.format("parquet")
            .mode('overwrite')
            .option("header", "true")
            .save("Datasets/airports_qa.parquet"))

# Salvando o arquivo em parquet

df7.write.parquet(
path = 'C:/Users/coskata/Downloads/Datasets/parquet/airports.parquet',
mode = 'overwrite'
)

# Teste do resultado

path = 'C:/Users/coskata/Downloads/Datasets/parquet/airports.parquet'
airports_parquet = spark.read.parquet(path)

airports_parquet.toPandas()

# Planes

# Pergunta 1

df1 = df_planes.withColumn('qa_tailnum', (
        F.when(check_empty_column('tailnum'), 'M')
         .when(~F.length(F.col('tailnum')).between(5, 6), 'S')
         .when(~F.col('tailnum').startswith('N'), 'FN')
         .when( F.col('tailnum').rlike(REGEX_NNUMBER_INVALID), 'FE')
         .when(~F.col('tailnum').rlike(REGEX_NNUMBER), 'F')))

# Teste do resultado

df1.groupBy('qa_tailnum').count().show()

# Pergunta 2

df2 = df1.withColumn(
    "qa_year",
    when(
            (
                (col('year').isNull()) |
                (col('year') == "NA")
            ), "M"
        )
    .when(col('year') < 1950, "I")
)

# Pergunta 2 revisada
- O código é o mesmo, alterando apenas a formatação. 


# Usando dataframe
df_planes = df_planes.withColumn('qa_year', (
        F.when(check_empty_column('year'), 'M')
         .when(F.col('year') < 1950, 'I')))

df_planes.show(5)

# Teste do resultado

df2.groupBy('qa_year').count().show()

# Pergunta 3

df3 = df2.withColumn(
    "qa_type", 
    when(
            (
                (col('type').isNull()) |
                (col('type') == "NA")
            ), "M"
        )
    .when(~col("type").rlike("[Fixed wing multi engine,\
                               Fixed wing single engine,\
                               Rotorcraft]"), "C")
)

## Pergunta 3 revisada
- mesmo código de forma diferente 

# Usando dataframe
PLANE_TYPES = ["Fixed wing multi engine", "Fixed wing single engine", "Rotorcraft"]
df_planes = df_planes.withColumn('qa_type', (
        F.when(check_empty_column('type'), 'M')
         .when(~F.col('type').isin(PLANE_TYPES), 'C')))

df_planes.show(5)

# Teste do resultado

df3.groupBy('qa_type').count().show()

# Pergunta 4

df4 = df3.withColumn(
    "qa_manufacturer",
    when(
            (
                (col('manufacturer').isNull()) |
                (col('manufacturer') == "NA")
            ), "M"
        )
    .when(~col("manufacturer").rlike("[AIRBUS,\
                                        BOEING,\
                                        BOMBARDIER,\
                                        CESSNA,\
                                        EMBRAER,\
                                        SIKORSKY,\
                                        CANADAIR,\
                                        PIPER,\
                                        MCDONNELL DOUGLAS,\
                                        CIRRUS,\
                                        BELL,\
                                        KILDALL GARY,\
                                        LAMBERT RICHARD,\
                                        BARKER JACK,\
                                        ROBINSON HELICOPTER,\
                                        GULFSTREAM,\
                                        MARZ BARRY]"), "C"))

# Pergunta 4 revisada 
- De forma diferente utilizando expressao lambda e rlike




# Usando dataframe
PLANES_MANUFACTURERS = ["AIRBUS","BOEING","BOMBARDIER","CESSNA","EMBRAER","SIKORSKY","CANADAIR","PIPER","MCDONNELL DOUGLAS","CIRRUS","BELL","KILDALL GARY","LAMBERT RICHARD","BARKER JACK","ROBINSON HELICOPTER","GULFSTREAM","MARZ BARRY"]
REGEX_MANUFACTURERS = r'|'.join(map(lambda x : f".*({x}).*", PLANES_MANUFACTURERS))

df_planes = df_planes.withColumn('qa_manufacturer', (
        F.when(check_empty_column('manufacturer'), 'M')
         .when(~F.col('manufacturer').rlike(REGEX_MANUFACTURERS), 'C')))

df_planes.show(3,vertical=True)

# Teste do resultado

df4.groupBy('qa_manufacturer').count().show()

# Pergunta 5

df5 = df4.withColumn(
    "qa_model", 
    when(
            (
                (col('model').isNull()) |
                (col('model') == "NA")
            ), "M"
        )
    .when(
            (
                (col('manufacturer') == "AIRBUS") & (~col('model').startswith("A")) |
                (col('manufacturer') == "BOEING") & (~col('model').startswith("7")) |
                (col('manufacturer') == "BOMBARDIER") & (~col('model').startswith("CL")) |
                (col('manufacturer') == "MCDONNELL DOUGLAS") & (~col('model').startswith("MD") | col('model').startswith("DC"))
            ), "F"
        )
)

# Pergunta 5 revisada
- Mesmo código, formatação diferente



# Usando dataframe
df_planes = df_planes.withColumn('qa_model', (
        F.when(check_empty_column('model'), 'M')
         .when(
             ((F.col('manufacturer').rlike(r'.*AIRBUS.*') & ~F.col('model').startswith('A')) |
              (F.col('manufacturer').rlike(r'.*BOEING.*') & ~F.col('model').startswith('7')) |
              (F.col('manufacturer').rlike(r'.*(BOMBARDIER|CANADAIR).*') & ~F.col('model').startswith('CL')) |
              (F.col('manufacturer').rlike(r'.*MCDONNELL DOUGLAS.*') &  ~(F.col('model').startswith('MD') | F.col('model').startswith('DC')) )), 'F')))

df_planes.show(5)

# Teste do resultado

df_planes.groupBy('qa_model').count().show()

# Pergunta 6

df6 = df5.withColumn(
    "qa_engines",
    when(
            (
                (col('engines').isNull()) |
                (col('engines') == "NA")
            ), "M"
        )
    .when(~col('engines').between(1,4), "I")
    .when(col('engines').rlike("^[0-9A-Z^]*$"), "A")
)

### Pergunta 6 revisada
- Mesmo código, formatação diferente

df_planes = df_planes.withColumn('qa_engines', (
        F.when(check_empty_column('engines'), 'M')
         .when(F.col("engines").between(1,4),'I')
         .when(F.col("engines").rlike(REGEX_ALPHA),'A')
))






df_planes.show(2,vertical=True)

# Teste do resultado

df_planes.groupBy('qa_engines').count().show()

# Pergunta 7

df7 = df6.withColumn(
    "qa_seats",
    when(
            (
                (col('seats').isNull()) |
                (col('seats') == "NA")
            ), "M"
        )
    .when(~col('seats').between(2,500), "I")
    .when(col('seats').rlike("^[0-9A-Z^]*$"), "A")
)

# Teste do resultado

df7.groupBy('qa_seats').count().show()

# Pergunta 8

df8 = df7.withColumn(
    "qa_speed",
    when(
            (
                (col('speed').isNull()) |
                (col('speed') == "NA")
            ), "M"
        )
    .when(~col('speed').between(50,150), "I")
    .when(col('speed').rlike("^[0-9A-Z^]*$"), "A")
)

# pergunta 8 revisada
- Mesmo codigo, formatacao diferente

df_planes = df_planes.withColumn('qa_speed', (
        F.when(check_empty_column('speed'), 'M')
         .when(F.col('speed').between(50, 150), 'I')
         .when(F.col('speed').rlike(REGEX_ALPHANUM), 'A')
))

df_planes.show(2,vertical=True)

# Teste do resultado

df_planes.groupBy('qa_speed').count().show()

# Pergunta 9

df9 = df8.withColumn(
    "qa_engine",
    when(
            (
                (col('engine').isNull()) |
                (col('engine') == "NA")
            ), "M"
        )
    .when(
            (
                (~col('engine').rlike("Turbo-fan")) &
                (~col('engine').rlike("Turbo-jet")) &
                (~col('engine').rlike("Turbo-prop")) &
                (~col('engine').rlike("Turbo-shaft")) &
                (~col('engine').rlike("4 CycleY"))
            ), "C"
         )
)

# Pergunta 9 revisada
- Mesmo código formatação diferente

nome = ["Turbo-fan","Turbo-jet","Turbo-prop","Turbo-shaft","4 Cycle"]

df_planes = df_planes.withColumn('qa_engine', (
        F.when(check_empty_column('engine'), 'M')
         .when(~F.col('engine').isin(nome), 'C')))

df_planes.show(2,vertical=True)

# Teste do resultado

df_planes.groupBy('qa_engine').count().show()

# Salvando o arquivo em parquet com as colunas qa e informações importantes

df9.select(
    col('tailnum'), 
    col('qa_tailnum'),
    col('qa_year'), 
    col('qa_type'), 
    col('qa_manufacturer'), 
    col('qa_model'), 
    col('qa_engines'), 
    col('qa_seats'), 
    col('qa_speed'), 
    col('qa_engine')

).write.parquet(
    path = 'C:/Users/coskata/Downloads/Datasets/parquet/planes.parquet',
    mode = 'overwrite'
)

# Salvando de outra maneira sem passar todas as colunas


(df_airports.select(df_airports.colRegex("`^qa_.*`"))
            .repartition(1) # coalesce
            .write.format("parquet")
            .mode('overwrite')
            .option("header", "true")
            .save("Datasets/airports_qa.parquet"))

# Salvando o arquivo em parquet

df9.write.parquet(
path = 'C:/Users/coskata/Downloads/Datasets/parquet/planes.parquet',
mode = 'overwrite'
)

# Teste do resultado

path = 'C:/Users/coskata/Downloads/Datasets/parquet/planes.parquet'
planes_parquet = spark.read.parquet(path)

planes_parquet.toPandas()

# Airports

# Pergunta 1

df1 = df_airports.withColumn(
    "qa_year_month_day", 
    when(
            (
                (col('year').isNull()) |
                (col('year') == "NA")
            ), "MY"
        )
    .when(
            (
                (col('month').isNull()) |
                (col('month') == "NA")
            ), "MM"
        )
    .when(
            (
                (col('day').isNull()) |
                (col('day') == "NA")
            ), "MD"
        )
    .when(col("year") < 1950, "IY")
    .when(~col("month").between(1,12), "IM")
    .when(
            (~col('day').between(1,31)) | 
            (col("month") == 2) & 
            (~col('day').between(1,29)), "ID"
        )
   )

# Pergunta 1 revisada
-- mesmo código, formatado de forma diferente

df_flights = df_flights.withColumn('qa_year_month_day',(
        F.when(check_empty_column('year'),'MY')
         .when(check_empty_column('month'),'MM')
         .when(check_empty_column('day'),'MD')
         .when((F.col('year') > 1950),'IY')
         .when((F.col('month').between(1,12)),'IM')
         .when((F.col('day').between(1,31)) &(F.col('day') =='2'),'ID')
         ))

# Teste do resultado

df_flights.groupBy('qa_year_month_day').count().show()

# Pergunta 2

df2 = df1.withColumn(
    "qa_hour_minute",
    when(
            (
                (col('hour').isNull()) |
                (col('hour') == "NA")
            ), "MH"
        )
    .when(
            (
                (col('minute').isNull()) |
                (col('minute') == "NA")
            ), "MM"
        )
    .when(~col('hour').between(0,24), "IH")
    .when(~col('minute').between(0, 59), "IM")
)

# Teste do resultado

df2.groupBy('qa_hour_minute').count().show()

# Pergunta 3

# Usando dataframe
df3 = df2.withColumn('qa_dep_arr_time', (
        F.when(check_empty_column('dep_time'), 'MD')
         .when(check_empty_column('arr_time'), 'MA')
         .when(F.col('dep_time').rlike(REGEX_TIME_FMT), 'FD')
         .when(F.col('arr_time').rlike(REGEX_TIME_FMT), 'FA')))

# Teste do resultado

df3.groupBy('qa_dep_arr_time').count().show()

# Pergunta 4

df4 = df3.withColumn(
    "qa_dep_arr_delay",
    when(
            (
                (col('dep_delay').isNull()) |
                (col('dep_delay') == "NA")
            ), "MD"
        )
    .when(
            (
                (col('arr_delay').isNull()) |
                (col('arr_delay') == "NA")
            ), "MA"
        )
)

# Pergunta 4 revisada
- mesmo código formatacao diferente

df_flights = df_flights.withColumn('qa_dep_arr_delay', (
        F.when((F.col('dep_delay').isNull()) |
               (F.col('dep_delay') == ''), 'MD')
        .when((F.col('arr_delay').isNull())  |
              (F.col('arr_delay') == ''), 'MA')
        
))

# Teste do resultado

df_flights.groupBy('qa_dep_arr_delay').count().show()

# Pergunta 5

df5 = df4.withColumn(
    "qa_carrier", 
    when(
            (
                (col('carrier').isNull()) |
                (col('carrier') == "NA")
            ), "M"
        )
    .when(~col('carrier').rlike("{2}(^[0-9][A-Z]|^[A-Z][0-9])$"), "F")
)

# Teste do resultado

df5.groupBy('qa_carrier').count().show()

# Pergunta 6

df6 = df5.withColumn('qa_tailnum', (
        F.when(check_empty_column('tailnum'), 'M')
         .when(~F.length(F.col('tailnum')).between(5, 6), 'S')
         .when(~F.col('tailnum').startswith('N'), 'FN')
         .when( F.col('tailnum').rlike(REGEX_NNUMBER_INVALID), 'FE')
         .when(~F.col('tailnum').rlike(REGEX_NNUMBER), 'F')))

# Teste do resultado

df6.groupBy('qa_tailnum').count().show()

# Pergunta 7

df7 = df6.withColumn(
    "qa_flight", 
    when(
            (
                (col('flight').isNull()) |
                (col('flight') == "NA")
            ), "M"
        )
    .when(~col("flight").rlike("[0-9]{4}$"), "F")
)

# Pergunta 7 Revisada
- mesmo código formatacao diferente

df_flights = df_flights.withColumn ('qa_flight',   (
    F.when(
         (F.col('flight').isNull()) | 
         (F.col('flight') == ''), 'M')
     .when(
         (~F.col('flight').rlike("[0-9]{4}")),'F')
    
))

# Teste do resultado

df_flights.groupBy(col('qa_flight')).count().show()

# Pergunta 8

df8 = df7.withColumn(
    "qa_origin_dest", 
    when(
            (
                (col('origin').isNull()) |
                (col('origin') == "NA")
            ), "MO"
        )
    .when(
            (
                (col('dest').isNull()) |
                (col('dest') == "NA")
            ), "MD"
        )
    .when(
            (
                (col('origin').rlike("^[A-Z]*")) |
                (col('origin').rlike("^[0-9]*"))
            ), "FO"
        )
    .when(
            (
                (col('dest').rlike("^[A-Z]*")) |
                (col('dest').rlike("^[0-9]*"))
            ), "FD"
        )
)

### pergunta 8 revisada
- mesmo código formatacao diferente

df_flights = df_flights.withColumn('qa_origin_dest', (
    F.when((F.col('origin').isNull()) | (F.col('origin') == ''), 'MO')
     .when((F.col('dest').isNull())| (F.col('dest') == ''), 'MD')
     .when((F.col('origin').rlike("[0-9a-zA-Z]{3}")),'FO')                            
     .when((F.col('dest').rlike("[0-9a-zA-Z]{3}")),'FD')
    
))

# Teste do resultado

df8.groupBy('qa_origin_dest').count().show()

# Pergunta 9

df9 = df8.withColumn(
    "qa_air_time", 
    when(
            (
                (col('air_time').isNull()) |
                (col('air_time') == "NA")
            ), "M"
        )
    .when(~col("air_time").between(20,500), "I")
)

# Teste do resultado

df9.groupBy('qa_air_time').count().show()

# Pergunta 10

df10 = df9.withColumn(
    "qa_distance", 
    when(
            (
                (col('distance').isNull()) |
                (col('distance') == "NA")
            ), "M"
        )
    .when(~col("distance").between(50,3000), "I")
)

# Teste do resultado

df10.groupBy('qa_distance').count().show()

# Pergunta 11

df11 = df10.withColumn('qa_distance_airtime', (
        F.when(check_empty_column('distance'), 'M')
         .when((F.col('air_time') >= F.col('distance') * 0.1 + 30), 'TL')
         .when((F.col('air_time') <= F.col('distance') * 0.1 + 10), 'TS')
         .otherwise('TR')))

# Teste do resultado

df11.groupBy('qa_distance_airtime').count().show()

# Salvando o arquivo em parquet com as colunas qa e informações importantes

df11.select(
    col('tailnum'), 
    col('flight'), 
    col('origin'), 
    col('dest'), 
    col('distance'), 
    col('qa_year_month_day'), 
    col('qa_hour_minute'), 
    col('qa_dep_arr_time'), 
    col('qa_dep_arr_delay'),
    col('qa_carrier'), 
    col('qa_tailnum'), 
    col('qa_flight'), 
    col('qa_origin_dest'), 
    col('qa_air_time'), 
    col('qa_distance_airtime')
).write.parquet(
    path = 'C:/Users/coskata/Downloads/Datasets/parquet/flights.parquet',
    mode = 'overwrite'
)

# Salvando o arquivo em parquet

df11.write.parquet(
path = 'C:/Users/coskata/Downloads/Datasets/parquet/flights.parquet',
mode = 'overwrite'
)

# Teste do resultado

path = 'C:/Users/coskata/Downloads/Datasets/parquet/flights.parquet'
flights_parquet = spark.read.parquet(path)

flights_parquet.toPandas()
